Add files via upload
It's time to let your imagination run  with the creation of a fun and interactive text-based Hangman game! You will be the designer in this project, and on players' wish, they may try to guess a word hidden by selecting one letter at a time. The game randomly chooses a word from a list and players are required to find it by making incorrect guesses within a given number of attempts.